+++ orix/quaternion/orientation.py
@@ -51,6 +51,7 @@ from orix.quaternion.orientation_region import OrientationRegion
 from orix.quaternion.rotation import Rotation
 from orix.quaternion.symmetry import C1, Symmetry
 from orix.scalar import Scalar
+from orix.vector import AxAngle
 from orix._util import deprecated
 
 
@@ -279,11 +280,10 @@ class Misorientation(Rotation):
         Examples
         --------
         >>> import numpy as np
-        >>> from orix.quaternion.symmetry import C4, C2
-        >>> from orix.quaternion.orientation import Misorientation
+        >>> from orix.quaternion import Misorientation, symmetry
         >>> data = np.array([[0.5, 0.5, 0.5, 0.5], [0, 1, 0, 0]])
         >>> m = Misorientation(data)
-        >>> m.symmetry = (C4, C2)
+        >>> m.symmetry = (symmetry.C4, symmetry.C2)
         >>> m = m.map_into_symmetry_reduced_zone()
         >>> m.distance()
         array([[3.14159265, 1.57079633],
@@ -506,8 +506,42 @@ class Orientation(Misorientation):
             o.symmetry = symmetry
         return o
 
+    @classmethod
+    def from_axes_angles(cls, axes, angles, symmetry=None):
+        """Creates orientation(s) from axis-angle pair(s).
+
+        Parameters
+        ----------
+        axes : Vector3d or array_like
+            The axis of rotation.
+        angles : array_like
+            The angle of rotation, in radians.
+        symmetry : Symmetry, optional
+            Symmetry of orientation(s). If None (default), no symmetry
+            is set.
+
+        Returns
+        -------
+        Orientation
+
+        Examples
+        --------
+        >>> import numpy as np
+        >>> from orix.quaternion import Orientation, symmetry
+        >>> ori = Orientation.from_axes_angles((0, 0, -1), np.pi / 2, symmetry.Oh)
+        >>> ori
+        Orientation (1,) m-3m
+        [[ 0.7071  0.      0.     -0.7071]]
+
+        See Also
+        --------
+        from_neo_euler
+        """
+        axangle = AxAngle.from_axes_angles(axes, angles)
+        return cls.from_neo_euler(axangle, symmetry)
+
     def angle_with(self, other):
-        """The symmetry reduced smallest angle of rotation transforming
+        """The smallest symmetry reduced angle of rotation transforming
         this orientation to the other.
 
         Parameters